requirements.txt
coverage==7.13.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
prometheus_client==0.23.1
PyJWT==2.10.1
psycopg2-binary
python-dotenv==1.2.1
